[games-menu] - Fix sort function

+++ applications/games_menu/modules/menu.c
@@ -1003,7 +1003,7 @@ static int AppCompareFirmwares(const void *a, const void *b)
 
 	if (ContainsOnlyNumbers(left->file) && ContainsOnlyNumbers(right->file))
 	{
-		return atol(left->file) - atol(right->file);
+		return atoi(left->file) - atoi(right->file);
 	}
 	else if (ContainsOnlyNumbers(left->file))
 	{
